fix: ensure we don't loop while reading pod logs
see https://github.com/jenkins-x-plugins/jx-build-controller/issues/30

we used to mark the pod as "completed" before reading its logs. which means that a running pod was never marked as "completed" until the 2nd loop, which means that the logs are seen as "duplicated".

the fix here is to check if the pod is completed after reading its logs - which should be the case, because we're waiting for each container.